requirements.txt
aiogram==2.22.2
asyncpg==0.26.0
SQLAlchemy==1.4.41
apscheduler==3.10.4
requests==2.32.3
PyJWT==2.9.0
